Actually correctly synced the health between host and client(s)
The previous fix for syncing health between the host and client was done to allow the client to change it's own health due to it not automatically changing using the new uNet networking. Now all health is controlled by the server and synced to the clients correctly so that the clients can't control health.