Added credits scene and fixed main menu bugs
Added a credits scene. It will need more logos in the future, but for now it's just a basic credits scene. There was also a bug in the main menu where you couldn't exit the game anymore. There was an extra object that got added to the heirarchy so that messed up the getChild(x) functions for the "Exit Game" button.