Fixed Main Menu bugs and added exit button
There were for some reason a lot of bugs in the main menu and particularly in the settings menu with many null references. Those are now fixed to allow the settings to be changed in-game again. Also added an "Exit" button to the Multiplayer lobby to allow the player to actually exit the lobby in order to exit the game without having to use Alt+F4.