Fixed notification bug
There was a bug with notifications when picking up wood. It would only show the notification on the server/host machine even if it wasn't the person that picked it up or on the same team. It now shows it on the correct screen and shows a notification for picking up wood, stone, and metal resources.